+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/w3c/JavascriptW3cPerformanceTimeline.kt
@@ -1,19 +1,25 @@
 package com.atlassian.performance.tools.jiraactions.api.w3c
 
-import com.atlassian.performance.tools.jiraactions.api.w3c.harvesters.getJsElementsPerformance
-import com.atlassian.performance.tools.jiraactions.api.w3c.harvesters.getJsNavigationsPerformance
-import com.atlassian.performance.tools.jiraactions.api.w3c.harvesters.getJsResourcesPerformance
+import com.atlassian.performance.tools.jiraactions.w3c.harvesters.getJsElementsPerformance
+import com.atlassian.performance.tools.jiraactions.w3c.harvesters.getJsNavigationsPerformance
+import com.atlassian.performance.tools.jiraactions.w3c.harvesters.getJsResourcesPerformance
 import org.openqa.selenium.JavascriptExecutor
 
 /**
  * Obtains entries from [javascript].
  */
-class JavascriptW3cPerformanceTimeline(
+class JavascriptW3cPerformanceTimeline @Deprecated("Use JavascriptW3cPerformanceTimeline.Builder instead.") constructor(
     javascript: JavascriptExecutor,
-    withNavigationPerformance: Boolean = true,
-    withResourcesPerformance: Boolean = true,
-    withElementPerformance: Boolean = true
+    withNavigationPerformance: Boolean,
+    withResourcesPerformance: Boolean,
+    withElementPerformance: Boolean
 ) : W3cPerformanceTimeline {
+
+    @Deprecated("Use JavascriptW3cPerformanceTimeline.Builder instead.")
+    constructor(
+        javascript: JavascriptExecutor
+    ) : this(javascript, true, true, true)
+
     private var getNavigationPerformance: () -> List<PerformanceNavigationTiming> =
         if (withNavigationPerformance) ({ getJsNavigationsPerformance(javascript) }) else ({ emptyList() })
     private var getResourcesPerformance: () -> List<PerformanceResourceTiming> =
@@ -26,6 +32,28 @@ class JavascriptW3cPerformanceTimeline(
             navigations = getNavigationPerformance(),
             resources = getResourcesPerformance(),
             elements = getElementPerformance()
+        )
+    }
+
+    class Builder(private val javascript: JavascriptExecutor) {
+        private var navigationPerformance = false
+        private var resourcesPerformance = false
+        private var elementPerformance = false
+
+        fun withNavigationPerformance() = apply { navigationPerformance = true }
+        fun withResourcesPerformance() = apply { resourcesPerformance = true }
+        fun withElementPerformance() = apply { elementPerformance = true }
+        fun withAllMetrics() = apply {
+            navigationPerformance = true
+            resourcesPerformance = true
+            elementPerformance = true
+        }
+
+        fun build() = JavascriptW3cPerformanceTimeline(
+            javascript = javascript,
+            withNavigationPerformance = navigationPerformance,
+            withResourcesPerformance = resourcesPerformance,
+            withElementPerformance = elementPerformance
         )
     }
 }