+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/w3c/RecordedPerformanceEntries.kt
@@ -7,5 +7,5 @@ package com.atlassian.performance.tools.jiraactions.api.w3c
 class RecordedPerformanceEntries internal constructor(
     val navigations: List<PerformanceNavigationTiming>,
     val resources: List<PerformanceResourceTiming>,
-    val elements: List<PerformanceElementTiming>?
+    val elements: List<PerformanceElementTiming>
 )